First Unsuccessful Attempt at Getting post-contrained torque & dipole moment values from torqueRodDynamicEffector
Created new internal message struct in simulation/fswInterfaceMessages called torqueRodIntMsg that has constrained dipole moment and torque, updated torqueRodDynamicEffector to write to this message, but it doesn't work. Looks like the values are being assigned before the MaxDipoleMoment constraint is applied. Not critical, but would be nice to have the data.